[ STYLE ] 코드 정리
