+++ app/src/main/java/com/lottomate/lottomate/presentation/screen/map/model/StoreBottomSheetExpendedType.kt
@@ -1,0 +1,10 @@
+package com.lottomate.lottomate.presentation.screen.map.model
+
+/**
+ * 복권 판매점 BottomSheet 의 확장/축소 타입
+ */
+enum class StoreBottomSheetExpendedType {
+    COLLAPSED,
+    HALF,
+    FULL,
+}
