+++ app/src/main/res/drawable/icon_copy.xml
@@ -5,6 +5,6 @@
     android:viewportHeight="24">
   <path
       android:pathData="M7,4.002C7,2.897 7.895,2.002 9,2.002H18C19.105,2.002 20,2.897 20,4.002V16.002C20,17.106 19.105,18.002 18,18.002H17V20.002C17,21.106 16.105,22.002 15,22.002H6C4.895,22.002 4,21.106 4,20.002V8.002C4,6.897 4.895,6.002 6,6.002H7V4.002ZM9,6.002H15C16.105,6.002 17,6.897 17,8.002V16.002H18L18,4.002H9V6.002ZM15,8.002L6,8.002V20.002H15V8.002Z"
-      android:fillColor="#000000"
+      android:fillColor="#6B6B6B"
       android:fillType="evenOdd"/>
 </vector>
